requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.0
Flask==2.3.3
fonttools==4.43.0
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
multidict==6.0.4
numpy==1.26.0
openai==0.28.1
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
pluggy==1.3.0
pyparsing==3.1.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
seaborn==0.13.0
six==1.16.0
SpeechRecognition==3.10.0
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.5
Werkzeug==2.3.7
yarl==1.9.2
